requirements.txt
torch
torchvision
numpy
matplotlib
scikit-learn
pytest
pytest-cov
pylint
build
black
jupyter
pillow
transformers
seaborn
tqdm
poetry